Fixed some bugs and added some of the functionality for receipt scoring

+++ Receipt Processor/TypeDefs.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef unsigned int uint;
+typedef uint16_t uint;
